refactor(indexer): split REST routes into per-endpoint files
The `routes/` files are named after their URL path (`graphql.rs` -> /graphql,
`ws.rs` -> /ws, `blocks.rs` -> the /block/* resource). `/query`, `/simulate`,
and `/broadcast` share no prefix or resource, so the transport-named `rest.rs`
was the odd one out. Split it into `query.rs`, `simulate.rs`, and
`broadcast.rs` (one handler each, registered individually like the `index`
routes), dropping the now-pointless empty `web::scope("")`.

No behavior change.
